+++ src/KubeLibrary/KubeLibrary.py
@@ -234,8 +234,6 @@ class KubeLibrary:
         Environment variables:
         - HTTP_PROXY:
           Proxy URL
-        - NO_PROXY:
-          Bypass proxy for host in the no_proxy list.
         """
         self.api_client = None
         self.cert_validation = cert_validation
@@ -263,8 +261,7 @@ class KubeLibrary:
         if not self.api_client:
             self.api_client = client.ApiClient(configuration=client.Configuration().get_default_copy())
 
-        self.api_client.configuration.proxy = environ.get('HTTP_PROXY')
-        self.api_client.configuration.no_proxy = environ.get('NO_PROXY')
+        self.api_client.configuration.proxy = environ.get('http_proxy') or environ.get('HTTP_PROXY')
 
         self._add_api('v1', client.CoreV1Api)
         self._add_api('networkingv1api', client.NetworkingV1Api)